Permisos de expedientes

+++ VIFAC/records/views.py
@@ -14,15 +14,19 @@ import datetime
 from .models import Expediente, Documento
 from .forms import RecordForm, DocumentForm
 
+from rolepermissions.checkers import has_role
+
 # Create your views here.
 
 def index(request):
-    context = {}
-    return render(request, 'records/index.html', context)
+    if has_role(request.user, ['expedientes', 'admin']):
+        context = {}
+        return render(request, 'records/index.html', context)
+    else:
+        raise Http404
 
 # Detail View
 class RecordDetailView(DetailView):
-
     model = Expediente
 
     def get_context_data(self, **kwargs):
@@ -32,44 +36,53 @@ class RecordDetailView(DetailView):
 # New DB entries
 
 def new_record(request):
-    context = {'today': datetime.datetime.now()}
-    
-    if request.method == "POST":
+    if has_role(request.user, ['expedientes', 'admin']):
+        context = {'today': datetime.datetime.now()}
         
-        # Get the form through POST
-        new_record_form = RecordForm(request.POST)
+        if request.method == "POST":
+            
+            # Get the form through POST
+            new_record_form = RecordForm(request.POST)
+            
+            # Validate form data
+            if new_record_form.is_valid():
+                # Get form variables
+                # Create donor object
+                context['record'] = Expediente.objects.create(**new_record_form.cleaned_data)
+                return list_record(request)
+            
+            context['form'] = new_record_form
+            return render(request, 'records/new_record.html', context)
         
-        # Validate form data
-        if new_record_form.is_valid():
-            # Get form variables
-            # Create donor object
-            context['record'] = Expediente.objects.create(**new_record_form.cleaned_data)
-            return list_record(request)
+        else:
+            new_record_form = RecordForm()
         
         context['form'] = new_record_form
+        
         return render(request, 'records/new_record.html', context)
-    
     else:
-        new_record_form = RecordForm()
-    
-    context['form'] = new_record_form
-    
-    return render(request, 'records/new_record.html', context)
+        raise Http404
 
 # List all records
 
 def list_record(request):
-    records = Expediente.objects.all().values()
-
-    return render(request, 'records/list_records.html', {'records': records})
+    if has_role(request.user, ['expedientes', 'admin']):
+        records = Expediente.objects.all().values()
+    
+        return render(request, 'records/list_records.html', {'records': records})
+    else:
+        raise Http404
 
 #Delete from DB
 
 def RecordDelete(request, pk):
-    model = get_object_or_404(Expediente, pk=pk)
-    # noinspection PyArgumentList
-    model.delete()
-    return HttpResponseRedirect(reverse('records:list_records'))
+    if has_role(request.user, ['expedientes', 'admin']):
+        model = get_object_or_404(Expediente, pk=pk)
+        # noinspection PyArgumentList
+        model.delete()
+        return HttpResponseRedirect(reverse('records:list_records'))
+    else:
+        raise Http404
 
 class RecordUpdate(UpdateView):
     model = Expediente
@@ -220,44 +233,56 @@ class RecordUpdate(UpdateView):
 
 
 def model_form_upload(request, exp_id):
-    exp = get_object_or_404(Expediente, pk=exp_id)
-    if request.method == 'POST':
-        form = DocumentForm(request.POST, request.FILES)
-        if form.is_valid():
-            documento = form.save(commit=False)
-            documento.expediente = exp
-            documento.save()
-            return HttpResponseRedirect(reverse('records:list_records'))
+    if has_role(request.user, ['expedientes', 'admin']):
+        exp = get_object_or_404(Expediente, pk=exp_id)
+        if request.method == 'POST':
+            form = DocumentForm(request.POST, request.FILES)
+            if form.is_valid():
+                documento = form.save(commit=False)
+                documento.expediente = exp
+                documento.save()
+                return HttpResponseRedirect(reverse('records:list_records'))
+        else:
+            form = DocumentForm()
+        return render(request, 'records/model_form_upload.html', {
+            'form': form,
+            'expediente': exp
+        })
     else:
-        form = DocumentForm()
-    return render(request, 'records/model_form_upload.html', {
-        'form': form,
-        'expediente': exp
-    })
+        raise Http404
 
 
 def list_documents(request, exp_id):
-    expediente = get_object_or_404(Expediente, pk=exp_id)
-    documentos = expediente.documento_set.all()
-    context = {
-        'expediente': expediente,
-        'documentos': documentos,
-    }
-    return render(request, 'records/document_list.html', context)
+    if has_role(request.user, ['expedientes', 'admin']):
+        expediente = get_object_or_404(Expediente, pk=exp_id)
+        documentos = expediente.documento_set.all()
+        context = {
+            'expediente': expediente,
+            'documentos': documentos,
+        }
+        return render(request, 'records/document_list.html', context)
+    else:
+        raise Http404
 
 
 def download(request, path):
-    file_path = os.path.join(settings.MEDIA_ROOT, path)
-    if os.path.exists(file_path):
-        with open(file_path, 'rb') as fh:
-            response = HttpResponse(fh.read(), content_type="application/vnd.ms-excel")
-            response['Content-Disposition'] = 'inline; filename=' + os.path.basename(file_path)
-            return response
+    if has_role(request.user, ['expedientes', 'admin']):
+        file_path = os.path.join(settings.MEDIA_ROOT, path)
+        if os.path.exists(file_path):
+            with open(file_path, 'rb') as fh:
+                response = HttpResponse(fh.read(), content_type="application/vnd.ms-excel")
+                response['Content-Disposition'] = 'inline; filename=' + os.path.basename(file_path)
+                return response
+        else:
+            raise Http404
     else:
         raise Http404
 
 
 def delete_document(request, doc_id):
-    document = Documento.objects.get(pk=doc_id)
-    document.delete()
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    if has_role(request.user, ['expedientes', 'admin']):
+        document = Documento.objects.get(pk=doc_id)
+        document.delete()
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    else:
+        raise Http404
